Altered initialize method so that R can call new("designMatrix")
R has changed and now if you want to use initialize you must be
careful that any specified parameters are also defined and defaulted
such that the object can be made by a call to new().  This is
necessary because the object has to be able to be made by new() in
order to be extendable.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/ArrayTools@38685 bc3139a8-67e5-0310-9ffc-ced21a209358